feat: move LoomianSetDto to LoomianSetRequestDto
